requirements.txt
# ── Training ───────────────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0
Pillow>=10.0.0

# ── Backend / API ──────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ── Utilities ──────────────────────────────────────────────────────────────────
aiofiles>=23.2.0
